feat(style): clean up settings menu styling
